Phase 2 (Todo Full Stack)/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pyjwt[cryptography]
asyncpg
pydantic-settings
python-dotenv
passlib[bcrypt]
python-multipart
